Added a secondary image size reduction to reduce false positives
The code 'numberplatev2.m' is the latest version, it has the same
functionality as normxcorrtest.m plus it resizes the template 8 times.
This is to find the size that has the highest peak, which is then used
to reduce the image size to only areas that have some success with the
template convolution.