Project 1 completed
Version 4 of numberplate was created and is partially successful,
however to date the most successful version is Version 2.  Version 3 has
edge detection implemented on top of version 2's binary convolution in
an attempt to get multilayered detection, however the edge detection was
inconsistent.  Version 4 consisted of implementing OCR in place of edge
detection, however this was found to be less accurate than anticipated.